requirements.txt
PyJWT==2.8.0
pymysql==1.1.0
boto3 ==1.28.78
stripe>=5.0.0